feat: add things, refactor state
