+++ src/navigation/index.ts
@@ -1,5 +1,3 @@
-import actionCreators from "./actions";
-import reducers from "./reducers";
 import ScreenNavigator, { registeredScreens } from "./navigator";
 
-export { reducers, actionCreators, registeredScreens, ScreenNavigator };
+export { registeredScreens, ScreenNavigator };
