+++ src/navigation/search/locales.ts
