+++ src/navigation/select-result/locales.ts
